modify float type columns to decimal type

+++ db/migrate/20220627094754_add_money_to_users.rb
@@ -1,5 +1,5 @@
 class AddMoneyToUsers < ActiveRecord::Migration[6.1]
   def change
-    add_column :users, :money, :float, default: 10000
+    add_column :users, :money, :decimal, default: 10000
   end
 end
